Osa_4 finished

+++ Osa_4/Backend/controllers/blogs.js
@@ -1,37 +1,61 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const logger = require('../utils/logger')
-const middleware = require('../utils/middleware')
+const {userExtractor} = require('../utils/middleware')
 
-
-
+//get all blogs
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
-
 })
 
-blogsRouter.post('/',  async (request, response) => {
-  //middleware.validateInput make my test long. So thats why not used here, even tho makes code prettier :)
-  if (!request.body.title || !request.body.url) {
-    return response.status(400).json({ error: 'Both, title and url are required' });
+//add new blog with authorization
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
+ 
+  const user = request.user
+  //Make sure title and url exists
+  if (!body.title || !body.url) {
+    return response.status(400).json({ error: 'Both, title and url are required' })
   }
-  const blog = new Blog(request.body)
+  //Create new blog
+  const blog = new Blog({
+    author: body.author,
+    url: body.url,
+    title: body.title,
+    likes: body.likes,
+    user: user._id
+  })
 
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
+  const savedBlog = await blog.save()
+  //add blog to User data
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+
+//Delete blog with authorization
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const userid = request.user._id
+
+  const blog = await Blog.findById(request.params.id)
+
+  //authored user must be the same as user that added the blog
+  if (blog.user.toString() === userid.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+  }else{
+    response.status(403).json({error: 'you can only delete your own blogs'})
+  }
   response.status(204).end()
 })
 
+//Modify likes of a blog. Authorization not included
 blogsRouter.put('/:id', async (request, response) => {
-  const {likes} = request.body
-  
-  const modifiedBlog = await Blog.findByIdAndUpdate(request.params.id, {likes}, { new: true })
+  const { likes } = request.body
+
+  const modifiedBlog = await Blog.findByIdAndUpdate(request.params.id, { likes }, { new: true })
   response.json(modifiedBlog)
 })
